pyproject.toml
[tool.poetry]
name = "dolphin-bot"
version = "0.1.0"
description = "Cognitive Computations: Dolphin bot"
authors = ["pabl-o-ce <cye@poscye.com>"]
license = "MIT"
readme = "README.md"

[[tool.poetry.source]]
name = "llama-cpp-python-wheels"
url = "https://jllllll.github.io/llama-cpp-python-cuBLAS-wheels/AVX512/cpu"

[tool.poetry.dependencies]
python = "^3.11"
discord-py-interactions = "^5.11.0"
python-dotenv = "^1.0.0"
llama-cpp-python = {version = ">=0.2.26", source = "llama-cpp-python-wheels"}
langchain = "^0.1.0"
langchain-community = "0.0.12"
langchain-core = "^0.1.8"
redis = "^5.0.1"
openai = "^1.8.0"
langchain-openai = "^0.0.2.post1"
discord-py = "^2.3.2"

[tool.poetry.scripts]
start-bot = "main:main"

[tool.poetry.group.dev.dependencies]
pylint = "^3.0.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
